Extract common logic from csv reader to base reader
